Integrate bower.js
https://github.com/bower/bower

Bower is a package manager for the web. It offers a generic,
unopinionated solution to the problem of front-end package management,
while exposing the package dependency model via an API that can be
consumed by a more opinionated build stack. And install a couple of
packages:

  * Underscore.js
  * Backbone.js
  * JQuery

This what I want to use in this project. My goal is to refactor this
project (spaghetti code) and use Backbone.js (structured
code). Integrate bower.js

And install a couple of packages:

    * Underscore.js
    * Backbone.js
    * JQuery

This what I want to use in this project. My goal is to refactor this
project (spaghetti code) and use Backbone.js (structured code).